refactor decompress and add invoke
