+++ provider/aliyun/invokefunction.go
@@ -4,11 +4,11 @@ import (
 	"github.com/aliyun/fc-go-sdk"
 )
 // InvokeFunction invokes a existed function
-func (m *Manager) InvokeFunction(funcName string, args *[]byte) ([]byte, error) {
+func (m *Manager) InvokeFunction(funcName string, args []byte) ([]byte, error) {
 	out, err := m.fcClient.InvokeFunction(&fc.InvokeFunctionInput{
 		ServiceName: &service,
 		FunctionName: &funcName,
-		Payload: args,
+		Payload: &args,
 	})
 	if err != nil {
 		return nil, err
